reducer separation changes
